requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==2.0.12
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.2
cymem==2.0.11
et_xmlfile==2.0.0
fastapi==0.116.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
mysql-connector-python==9.4.0
narwhals==2.6.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pillow==11.3.0
plotly==6.3.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pt_core_news_md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.8.0/pt_core_news_md-3.8.0-py3-none-any.whl#sha256=54382cda034e41f3ec605ceeb924cd6cfdbced00a65f7afa12218520ba7008c5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.27.1
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
starlette==0.47.3
streamlit==1.49.1
tenacity==9.1.2
thinc==8.3.6
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.3
